Fix several -Wconversion warnings in .hh files
Before this patch enabling the '-Wconversion' warning flag generated
~14k lines of output. Though many warnings triggered in header files and
thus are repeated several times. This patch fixes (most of) those
warnings and roughly halves the size of the output.

I'm still evaluating whether or not to enable the -Wconversion flag by
default (currently it's not). A good way to do this is to try to fix
those warnings.

So far, by fixing these warnings, I found one real bug in utf8_core.hh:
    bool is_surrogate(uint16_t cp)
This function took a 'uint16_t' parameter, but it should have been a
'uint32_t'. Meaning that code points like 0x1d800 would wrongly be
classified as a surrogate. That's a bug, but it's likely this never
caused problems in practice.

Maybe this single bug is already enough justification for trying to be
-Wconversion clean?

+++ src/commands/TclObject.hh
@@ -262,7 +262,7 @@ private:
 		if (objc == 0) return; // because 0-length VLAs are not allowed (but gcc/clang allow it as an extension)
 		VLA(Tcl_Obj*, objv, objc);
 		std::transform(first, last, objv.data(), [](const auto& t) { return newObj(t); });
-		addListElementsImpl(objc, objv.data());
+		addListElementsImpl(narrow<int>(objc), objv.data());
 	}
 
 	void addListElement(Tcl_Obj* element);
